version 2.0

+++ test_make.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
+
+if [ "$1" == "" ]; then
+    echo "Usage: $0 <file>"
+    exit 1
+fi
+FILE="$1"
+
 if [ "x${1}" == "xdist" ] ; then
   ARC="${PWD##*/}" ; rm -f office_analyzer
   tar cpzf "../${ARC}.tar.gz" .
@@ -6,4 +13,4 @@ if [ "x${1}" == "xdist" ] ; then
 fi
 if [ "x${1}" == "xclean" ] ; then rm -f office_analyzer ; exit ; fi
 gcc -s -o office_analyzer test.c -lncursesw -lzip
-./office_analyzer test.docx
+./office_analyzer "$FILE"
